FEAT : Validator 관련 코드 작성

+++ src/main/java/umc/study/repository/store/StoreRepositoryCustom.java
@@ -3,8 +3,10 @@ package umc.study.repository.store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.study.domain.store.Store;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface StoreRepositoryCustom {
-    List<Store> dynamicQueryWithBooleanBuilder(String name, Float score);
+    List<Store> dynamicQueryWithBooleanBuilder(String name, BigDecimal score);
+
 }
