+++ src/main/java/umc/study/service/store/StoreQueryService.java
@@ -2,6 +2,7 @@ package umc.study.service.store;
 
 import umc.study.domain.store.Store;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,5 +11,5 @@ public interface StoreQueryService {
     //아래 메서드는 Store가 존재하지 않을 수 있기에 Optional 사용
     Optional<Store> findStore(Long id);
 
-    List<Store> findStoresByNameAndScore(String name, Float score);
+    List<Store> findStoresByNameAndScore(String name, BigDecimal score);
 }
